OTV: Move BroadcastManager into package
