+++ cmd/oceantv/broadcast_manager.go
@@ -25,451 +25,3 @@ LICENSE
 */
 
 package main
-
-import (
-	"context"
-	"errors"
-	"fmt"
-	"strings"
-	"time"
-
-	rv_config "github.com/ausocean/av/revid/config"
-	"github.com/ausocean/cloud/cmd/oceantv/broadcast"
-	"github.com/ausocean/cloud/cmd/oceantv/manager"
-	"github.com/ausocean/cloud/cmd/oceantv/notifier"
-	"github.com/ausocean/cloud/cmd/oceantv/ratelimit"
-	"github.com/ausocean/cloud/cmd/oceantv/yt"
-	"github.com/ausocean/cloud/datastore"
-	"github.com/ausocean/cloud/model"
-	"github.com/ausocean/utils/nmea"
-	"github.com/google/uuid"
-)
-
-// OceanBroadcastManager is an implementation of BroadcastManager with
-// a particular focus around ocean broadcasts and AusOcean's infrastructure.
-type OceanBroadcastManager struct {
-	svc   Svc
-	log   func(string, ...interface{})
-	cfg   *Cfg
-	store Store
-}
-
-// newOceanBroadcastManager creates a new OceanBroadcastManager.
-// svc may be nil, but any methods that require it will panic.
-func newOceanBroadcastManager(svc Svc, cfg *Cfg, store Store, log func(string, ...interface{})) *OceanBroadcastManager {
-	return &OceanBroadcastManager{svc: svc, cfg: cfg, store: store, log: log}
-}
-
-func (m *OceanBroadcastManager) CreateBroadcast(
-	cfg *Cfg,
-	store Store,
-	svc Svc,
-) error {
-	// Only create a new broadcast if a valid one doesn't already exist.
-	if m.broadcastCanBeReused(cfg, svc) {
-		m.log("broadcast already exists with broadcastID: %s, streamID: %s", cfg.BID, cfg.SID)
-		err := m.Save(nil, func(_cfg *Cfg) {
-			_cfg.BID = cfg.BID
-			_cfg.SID = cfg.SID
-			_cfg.CID = cfg.CID
-			_cfg.RTMPKey = cfg.RTMPKey
-		})
-		if err != nil {
-			return fmt.Errorf("could not save broadcast config: %w", err)
-		}
-		return nil
-	}
-
-	// We're going to add the date to the broadcast's name, so get this and format.
-	loc, err := time.LoadLocation(locationID)
-	if err != nil {
-		return fmt.Errorf("could not load location: %w", err)
-	}
-
-	const layout = "02/01/2006"
-	dateStr := time.Now().In(loc).Format(layout)
-
-	const (
-		// This allows for 10 broadcasts to be created with 3 retries each
-		// all being started within the same hour.
-		limiterMaxTokens  = 30.0
-		limiterRefillRate = 2.0 // per hour
-		limiterID         = "ocean_token_bucket"
-	)
-	limiter, err := ratelimit.GetOceanTokenBucketLimiter(limiterMaxTokens, limiterRefillRate, limiterID, store)
-	if err != nil {
-		return fmt.Errorf("could not get token bucket limiter: %w", err)
-	}
-
-	timeCreated := time.Now().Add(1 * time.Minute)
-	resp, ids, rtmpKey, err := svc.CreateBroadcast(
-		context.Background(),
-		cfg.Name+" "+dateStr,
-		cfg.Description,
-		cfg.StreamName,
-		cfg.LivePrivacy,
-		cfg.Resolution,
-		timeCreated,
-		cfg.End,
-		yt.WithRateLimiter(limiter),
-	)
-	if err != nil {
-		return fmt.Errorf("could not create broadcast: %w, resp: %v", err, resp)
-	}
-	err = m.Save(nil, func(_cfg *Cfg) {
-		_cfg.BID = ids.BID
-		_cfg.SID = ids.SID
-		_cfg.CID = ids.CID
-		_cfg.RTMPKey = rtmpKey
-	})
-	if err != nil {
-		return fmt.Errorf("could not update config with transaction: %w", err)
-	}
-	return nil
-}
-
-// StartBroadcast starts a broadcast using the youtube live streaming API.
-// It provides AusOcean specific callbacks for starting and stopping
-// external hardware (cameras) and services.
-func (m *OceanBroadcastManager) StartBroadcast(
-	ctx Ctx,
-	cfg *Cfg,
-	store Store,
-	svc Svc,
-	extStart func() error,
-	onSuccess func(),
-	onFailure func(error),
-) {
-	if extStart != nil {
-		err := extStart()
-		if err != nil {
-			onFailure(fmt.Errorf("could not start external hardware: %w", err))
-			return
-		}
-	}
-
-	go func() {
-		err := svc.StartBroadcast(
-			cfg.Name,
-			cfg.BID,
-			cfg.SID,
-			saveLinkFunc(),
-			func() error { return nil }, // This is now handled by the hardware state machine.
-			func() error { return nil }, // This is now handled by the hardware state machine.
-			opsHealthNotifyFunc(ctx, cfg),
-			func() error { return nil }) // This is now handled by the hardware state machine.
-		if err != nil {
-			onFailure(fmt.Errorf("could not start broadcast: %w", err))
-			return
-		}
-		onSuccess()
-	}()
-}
-
-// StopBroadcast stops a broadcast using the youtube live streaming API.
-// It uses AusOcean methods for saving and stopping external hardware.
-func (m *OceanBroadcastManager) StopBroadcast(ctx Ctx, cfg *Cfg, store Store, svc Svc) error {
-	m.log("stopping broadcast")
-
-	status, err := svc.BroadcastStatus(ctx, cfg.BID)
-	if err != nil {
-		return fmt.Errorf("could not get broadcast status: %w", err)
-	}
-
-	if status != yt.StatusComplete && status != "" {
-		err := svc.CompleteBroadcast(ctx, cfg.BID)
-		if err != nil {
-			return fmt.Errorf("could not complete broadcast: %w", err)
-		}
-	}
-
-	err = m.Save(ctx, func(_cfg *Cfg) { _cfg.Active = false })
-	if err != nil {
-		return fmt.Errorf("could not save broadcast config, to update Active state: %w", err)
-	}
-
-	// Change privacy to post live privacy.
-	// This will also set the privacy of the video after the broadcast has ended.
-	err = svc.SetBroadcastPrivacy(ctx, cfg.BID, cfg.PostLivePrivacy)
-	if err != nil {
-		return fmt.Errorf("could not update broadcast privacy: %w", err)
-	}
-
-	return nil
-}
-
-// Save performs broadcast configuration update operations.
-// If ctx is nil, the background context will be used.
-//
-// update allows for the update of specific fields. After this update takes
-// place, the config we currently point at will be updated with any changes
-// that we have applied, and with anything from the store before the update.
-// If this is nil, the config currently in store will be replaced.
-func (m *OceanBroadcastManager) Save(ctx Ctx, update func(_cfg *Cfg)) error {
-	if ctx == nil {
-		ctx = context.Background()
-	}
-	_update := func(_cfg *Cfg) { *_cfg = *m.cfg }
-	if update != nil {
-		_update = func(_cfg *Cfg) {
-			update(_cfg)
-			*m.cfg = *_cfg
-		}
-	}
-
-	// Reference by UUID if we have one.
-	if uuid.Validate(m.cfg.UUID) == nil {
-		return broadcast.UpdateConfigWithTransaction(ctx, m.store, m.cfg.SKey, m.cfg.UUID, _update)
-	}
-
-	// If we don't have a UUID, we will follow the following steps:
-	//   1. Add a UUID to the config
-	//   2. Run the transaction to update the config
-	//   3. Copy the config, and index it by the UUID
-	//   4. Delete the old config
-	origUpdate := _update
-
-	_update = func(_cfg *Cfg) {
-		_cfg.UUID = uuid.NewString()
-		origUpdate(_cfg)
-	}
-
-	err := broadcast.UpdateConfigWithTransaction(ctx, m.store, m.cfg.SKey, m.cfg.Name, _update)
-	if err != nil {
-		return fmt.Errorf("unable to update config: %w", err)
-	}
-	oldName := broadcast.Scope + "." + m.cfg.Name
-	v, err := model.GetVariable(ctx, m.store, m.cfg.SKey, oldName)
-	if err != nil {
-		return fmt.Errorf("unable to get variable for updated config: %w", err)
-	}
-
-	newName := broadcast.Scope + "." + m.cfg.UUID
-	err = model.PutVariable(ctx, m.store, m.cfg.SKey, newName, v.Value)
-	if err != nil {
-		return fmt.Errorf("unable to put config indexed with UUID: %w", err)
-	}
-
-	err = model.DeleteVariable(ctx, m.store, m.cfg.SKey, oldName)
-	if err != nil {
-		return fmt.Errorf("unable to delete config indexed by name: %w", err)
-	}
-
-	return nil
-}
-
-// HandleStatus checks the status of a broadcast and stops it if it has
-// complete or revoked status.
-func (m *OceanBroadcastManager) HandleStatus(ctx Ctx, cfg *Cfg, store Store, svc Svc, noBroadcastCallBack manager.BroadcastCallback) error {
-	m.log("handling status check")
-	status, err := svc.BroadcastStatus(ctx, cfg.BID)
-	if err != nil {
-		if !errors.Is(err, yt.ErrNoBroadcastItems) {
-			return fmt.Errorf("could not get broadcast status: %w", err)
-		}
-
-		m.log("no broadcast items with this configuration listed")
-		err := noBroadcastCallBack(ctx, cfg, store, svc)
-		if err != nil {
-			return fmt.Errorf("could not call no broadcast callback: %w", err)
-		}
-	}
-
-	if status != yt.StatusComplete && status != yt.StatusRevoked {
-		return nil
-	}
-
-	m.log("status is complete or revoked")
-	err = noBroadcastCallBack(ctx, cfg, store, svc)
-	if err != nil {
-		return fmt.Errorf("could not call no broadcast callback: %w", err)
-	}
-
-	return nil
-}
-
-// HandleChatMessage generates a message with sensor readings for the
-// relevant site and posts the message to the broadcast chat. This works by
-// searching the site for any registered ESP devices and looking at the latest
-// signal values on sensors which have been marked true to send a message.
-func (m *OceanBroadcastManager) HandleChatMessage(ctx Ctx, cfg *Cfg) error {
-	if !cfg.SendMsg {
-		m.log("ignoring sensors")
-		return nil
-	}
-
-	m.log("building message")
-	var msg string
-
-	for _, sensor := range cfg.SensorList {
-		if !sensor.SendMsg {
-			continue
-		}
-		// Get the latest signal for the sensor.
-		var qty string
-
-		scalar, err := model.GetLatestScalar(ctx, m.store, model.ToSID(model.MacDecode(sensor.DeviceMac), sensor.Sensor.Pin))
-		if err == datastore.ErrNoSuchEntity {
-			continue
-		} else if err != nil {
-			return fmt.Errorf("could not get scalar for chat message: %v", err)
-		}
-
-		value, err := sensor.Sensor.Transform(scalar.Value)
-		if err != nil {
-			return fmt.Errorf("could not transform scalar: %v", err)
-		}
-
-		for _, q := range nmea.DefaultQuantities() {
-			if q.Code == nmea.Code(sensor.Sensor.Quantity) {
-				qty = q.Name
-			}
-		}
-
-		// Add the latest sensor value to the message.
-		var line string
-		if msg == "" {
-			line = fmt.Sprintf("%s: %3.1f %s ", qty, value, sensor.Sensor.Units)
-		} else {
-			line = fmt.Sprintf("| %s: %3.1f %s ", qty, value, sensor.Sensor.Units)
-		}
-		msg += line
-	}
-
-	if msg == "" {
-		m.log("chat message empty")
-		return nil
-	}
-
-	err := m.svc.PostChatMessage(cfg.CID, msg)
-	if err != nil {
-		return fmt.Errorf("broadcast chat message post error: %w", err)
-	}
-	return nil
-}
-
-// HandleHealth interprets the health of a broadcast and calls the provided callbacks in response to the health.
-// For tolerance to temporary issues, we only call the badHealthCallback if the health is bad for more than 4 checks.
-func (m *OceanBroadcastManager) HandleHealth(ctx Ctx, cfg *Cfg, store Store, goodHealthCallback func(), badHealthCallback func(string)) error {
-	m.log("handling health check")
-	issue, err := m.svc.BroadcastHealth(ctx, cfg.SID)
-	if err != nil {
-		return fmt.Errorf("could not check for stream issues: %w", err)
-	}
-
-	if issue == "" {
-		cfg.Issues = 0
-		goodHealthCallback()
-		return nil
-	}
-	m.log("issue found: %s", issue)
-
-	cfg.Issues++
-	const maxHealthIssues = 4
-	if cfg.Issues > maxHealthIssues {
-		badHealthCallback(issue)
-		cfg.Issues = 0
-	}
-
-	return m.Save(nil, func(_cfg *Cfg) { _cfg.Issues = cfg.Issues; *cfg = *_cfg })
-}
-
-func (m *OceanBroadcastManager) SetupSecondary(ctx Ctx, cfg *Cfg, store Store) error {
-	m.log("setting up vidforward broadcasting for %v", cfg.Name)
-
-	// Sanity check. This should only be invoked for the primary broadcast only so make sure
-	// the name does not contain the secondary broadcast postfix.
-	if strings.Contains(cfg.Name, broadcast.SecondaryPostfix) {
-		panic("setupVidforwardBroadcasting should only be invoked for the primary broadcast")
-	}
-
-	// Let's first set up the device.
-	// Set the HTTPAddress variable to send to the vidforward service.
-	// Set the Outputs variable to HTTP so that we're using MPEG-TS over HTTP.
-	mac := fmt.Sprintf("%012x", cfg.CameraMac)
-	err := setVar(ctx, store, mac+"."+rv_config.KeyHTTPAddress, cfg.VidforwardHost, cfg.SKey, m.log)
-	if err != nil {
-		return fmt.Errorf("could not set the HTTPAddress variable for the camera: %w", err)
-	}
-	err = setVar(ctx, store, mac+"."+rv_config.KeyOutputs, "HTTP", cfg.SKey, m.log)
-	if err != nil {
-		return fmt.Errorf("could not set the camera output to http: %w", err)
-	}
-	// Check if secondary broadcast already exists.
-	secondaryName := cfg.Name + broadcast.SecondaryPostfix
-
-	populateFields := func(_cfg *Cfg) {
-		// The secondary broadcast will for the most part copy the long term broadcast
-		// configuration, except for a few of the fields.
-		_cfg.Name = secondaryName
-		_cfg.StreamName = secondaryName
-		_cfg.LivePrivacy = "unlisted"     // We don't want the secondary broadcast to be easily discovered by youtube watchers.
-		_cfg.PostLivePrivacy = "unlisted" // This will be public eventually, but not while the software is young.
-		_cfg.OnActions = ""               // We don't need it to have any control of the camera hardware.
-		_cfg.OffActions = ""              // Ditto.
-		_cfg.SendMsg = true               // It would be handy to have sensors stored in the store broadcasts too.
-		_cfg.Start = cfg.Start
-		_cfg.End = cfg.End
-		_cfg.Resolution = cfg.Resolution
-		_cfg.Enabled = true
-	}
-
-	secondary, err := broadcastByName(cfg.SKey, secondaryName)
-	switch {
-	// Broadcast not found, so we need to create it.
-	case errors.Is(err, ErrBroadcastNotFound{}):
-		secondaryCfg := *cfg
-		populateFields(&secondaryCfg)
-
-		// Create a temporary OceanBroadcastManager for the secondary broadcast and create it (no update func required).
-		err = newOceanBroadcastManager(nil, &secondaryCfg, store, m.log).Save(ctx, nil)
-		if err != nil {
-			return fmt.Errorf("could not save secondary broadcast: %w", err)
-		}
-	case err != nil:
-		return fmt.Errorf("could not check if secondary broadcast exists: %w", err)
-
-	// Broadcast found so we need to update it with a transaction.
-	default:
-		// Create a temporary OceanBroadcastManager for the secondary broadcast and update it.
-		err = newOceanBroadcastManager(nil, secondary, store, m.log).Save(ctx, populateFields)
-		if err != nil {
-			return fmt.Errorf("could not update secondary broadcast: %w", err)
-		}
-	}
-
-	return nil
-}
-
-// opsHealthNotifyFunc returns a closure of notifier.Send given to the
-// broadcast.BroadcastStream function for notifiers.
-func opsHealthNotifyFunc(ctx Ctx, cfg *Cfg) func(string) error {
-	return func(msg string) error {
-		return notifier.N.Send(ctx, cfg.SKey, notifier.KindGeneric, msg)
-	}
-}
-
-// broadcastCanBeReused checks if a broadcast can be reused based on how old it
-// is, if it has been revoked or completed, and if its IDs have been set.
-func (m *OceanBroadcastManager) broadcastCanBeReused(cfg *Cfg, svc Svc) bool {
-	// Check if the broadcast was created today. Don't reuse an old broadcast.
-	startTime, err := svc.BroadcastScheduledStartTime(context.Background(), cfg.BID)
-	if err != nil {
-		m.log("could not get today's broadcast start time: %v", err)
-		return false
-	}
-	now := time.Now()
-	startOfToday := time.Date(now.Year(), now.Month(), now.Day(), 0, 0, 0, 0, now.Location())
-	if startTime.Before(startOfToday) || startTime.IsZero() {
-		m.log("broadcast does not exist for today, last start time: %v", startTime)
-		return false
-	}
-
-	status, err := svc.BroadcastStatus(context.Background(), cfg.BID)
-	if err != nil {
-		m.log("could not get today's broadcast status: %v", err)
-		return false
-	}
-	m.log("today's broadcast has status: %s", status)
-	return cfg.BID != "" && cfg.SID != "" && status != "" && status != yt.StatusRevoked && status != yt.StatusComplete
-}
