+++ cmd/oceantv/broadcast_test.go
@@ -44,26 +44,6 @@ import (
 	"github.com/ausocean/cloud/notify"
 )
 
-// TestRemoveDate tests the removeDate helper function.
-func TestRemoveDate(t *testing.T) {
-	tests := []struct {
-		in   string
-		want string
-	}{
-		{in: "A Broadcast 23/04/15", want: "A Broadcast "},
-		{in: "A Broadcast 04/23/15", want: "A Broadcast "},
-		{in: "ABroadcast04/23/15", want: "ABroadcast"},
-		{in: "ABroadcast04/23/15AStream", want: "ABroadcastAStream"},
-	}
-
-	for i, test := range tests {
-		got := removeDate(test.in)
-		if got != test.want {
-			t.Errorf("did not get expected result for test no. %d \ngot: %s \nwant: %s", i, got, test.want)
-		}
-	}
-}
-
 // dummyManager is a dummy implementation of the broadcastManager interface.
 type dummyManager struct {
 	cfg                                                                *Cfg
